requirements.txt
snowflake-ml-python>=1.22.0
snowflake-snowpark-python
scikit-learn
pandas
numpy
matplotlib
seaborn
joblib
pyarrow
